+++ ATSAccessibility/Reflection/InstitutionReflection.cs
@@ -343,5 +343,9 @@ namespace ATSAccessibility.Reflection {
 				return false;
 			}
 		}
+
+		public static int LogCacheStatus() {
+			return ReflectionValidator.TriggerAndValidate(typeof(InstitutionReflection), "InstitutionReflection");
+		}
 	}
 }