+++ ATSAccessibility/Reflection/RelicReflection.cs
@@ -1352,5 +1352,9 @@ namespace ATSAccessibility.Reflection {
 			}
 			return result;
 		}
+
+		public static int LogCacheStatus() {
+			return ReflectionValidator.TriggerAndValidate(typeof(RelicReflection), "RelicReflection");
+		}
 	}
 }